Handle files related info.
For example, the following HTML:

<input type="file" name="files[]" />
<input type="text" name="files[name][]" value="file 1" />
<input type="text" name="files[caption][]" value="" />

<input type="file" name="files[]" />
<input type="text" name="files[name][]" value="file 2" />
<input type="text" name="files[caption][]" value="caption 2" />

Will be translated into this:

Array
(
    [0] => Array
        (
            [grp] => files
            [info] => Array
                (
                    [name] => file 1
                    [caption] =>
                )

            [name] =>
            [type] =>
            [tmp_name] =>
            [error] => 4
            [size] => 0
        )

    [1] => Array
        (
            [grp] => files
            [info] => Array
                (
                    [name] => file 2
                    [caption] => caption 2
                )

            [name] => filename.extension
            [type] => image/png
            [tmp_name] => /tmp/phpU4TOpb
            [error] => 0
            [size] => 211189
        )

)

Important: info fields *must* be defined for all files otherwise they will not
match driving to data disaster.

This feature is very handy but if you for example forget to put the 1st caption
field, then the text "caption 2" will be inserted into the 1st element of the
array instead of the second. This may be a serious data mismatch so be careful.